refactor(web-search): consolidate lib into lib/webSearch/, drop MASKED_API_KEY_PLACEHOLDER
Moves all types, utils, svc, and hooks out of the WebSearchPage directory and
into lib/webSearch/ (mirrors lib/agents/ and lib/languageModels/ patterns).

Drops MASKED_API_KEY_PLACEHOLDER in favour of isNonRevealable={hasStoredKey} on
PasswordInputTypeInField — the stored-key state is now expressed via Formik
validation (api_key only required when requiresApiKey && !hasStoredKey) and the
apiKeyChanged flag (truthy only when user actually types a new value).
