--- web/src/refresh-pages/admin/WebSearchPage/searchProviderUtils.ts
@@ -1,266 +1,0 @@
-import type { IconFunctionComponent } from "@opal/types";
-import {
-  SvgBrave,
-  SvgExa,
-  SvgGoogle,
-  SvgSearxng,
-  SvgSerper,
-} from "@opal/logos";
-
-export type WebSearchProviderType =
-  | "google_pse"
-  | "serper"
-  | "exa"
-  | "searxng"
-  | "brave";
-
-export const SEARCH_PROVIDER_DETAILS: Record<
-  WebSearchProviderType,
-  {
-    label: string;
-    subtitle: string;
-    helper: string;
-    logo?: IconFunctionComponent;
-    apiKeyUrl?: string;
-  }
-> = {
-  exa: {
-    label: "Exa",
-    subtitle: "Exa.ai",
-    helper: "Connect to Exa to set up web search.",
-    logo: SvgExa,
-    apiKeyUrl: "https://dashboard.exa.ai/api-keys",
-  },
-  serper: {
-    label: "Serper",
-    subtitle: "Serper.dev",
-    helper: "Connect to Serper to set up web search.",
-    logo: SvgSerper,
-    apiKeyUrl: "https://serper.dev/api-key",
-  },
-  brave: {
-    label: "Brave",
-    subtitle: "Brave Search API",
-    helper: "Connect to Brave Search API to set up web search.",
-    logo: SvgBrave,
-    apiKeyUrl:
-      "https://api-dashboard.search.brave.com/app/documentation/web-search/get-started",
-  },
-  google_pse: {
-    label: "Google PSE",
-    subtitle: "Google",
-    helper: "Connect to Google PSE to set up web search.",
-    logo: SvgGoogle,
-    apiKeyUrl: "https://programmablesearchengine.google.com/controlpanel/all",
-  },
-  searxng: {
-    label: "SearXNG",
-    subtitle: "SearXNG",
-    helper: "Connect to SearXNG to set up web search.",
-    logo: SvgSearxng,
-  },
-};
-
-/**
- * Display order for built-in providers.
- * Derived from insertion order of `SEARCH_PROVIDER_DETAILS` for a single source of truth.
- */
-export const SEARCH_PROVIDER_ORDER = Object.keys(
-  SEARCH_PROVIDER_DETAILS
-) as WebSearchProviderType[];
-
-export function getSearchProviderDisplayLabel(
-  providerType: string,
-  providerName?: string | null
-): string {
-  if (providerName) return providerName;
-  return (
-    (SEARCH_PROVIDER_DETAILS as Record<string, { label: string }>)[providerType]
-      ?.label ?? providerType
-  );
-}
-
-export type SearchProviderConfig = Record<string, string> | null | undefined;
-
-export type SearchProviderLike =
-  | {
-      has_api_key: boolean;
-      config: SearchProviderConfig;
-    }
-  | null
-  | undefined;
-
-type SearchProviderCapabilities = {
-  requiresApiKey: boolean;
-  /** Keys required in `config` to consider the provider configured / connectable. */
-  requiredConfigKeys: string[];
-  /**
-   * Some providers historically stored config under different keys.
-   * When reading stored config, we consider these aliases equivalent.
-   */
-  storedConfigAliases?: Record<string, string[]>;
-};
-
-const SEARCH_PROVIDER_CAPABILITIES: Record<
-  WebSearchProviderType,
-  SearchProviderCapabilities
-> = {
-  exa: {
-    requiresApiKey: true,
-    requiredConfigKeys: [],
-  },
-  serper: {
-    requiresApiKey: true,
-    requiredConfigKeys: [],
-  },
-  brave: {
-    requiresApiKey: true,
-    requiredConfigKeys: [],
-  },
-  google_pse: {
-    requiresApiKey: true,
-    requiredConfigKeys: ["search_engine_id"],
-    storedConfigAliases: {
-      search_engine_id: ["search_engine_id", "cx", "search_engine"],
-    },
-  },
-  searxng: {
-    requiresApiKey: false,
-    requiredConfigKeys: ["searxng_base_url"],
-    storedConfigAliases: {
-      searxng_base_url: ["searxng_base_url"],
-    },
-  },
-};
-
-const DEFAULT_SEARCH_PROVIDER_CAPABILITIES: SearchProviderCapabilities = {
-  requiresApiKey: true,
-  requiredConfigKeys: [],
-};
-
-function getCapabilities(providerType: string): SearchProviderCapabilities {
-  return (
-    (
-      SEARCH_PROVIDER_CAPABILITIES as Record<string, SearchProviderCapabilities>
-    )[providerType] ?? DEFAULT_SEARCH_PROVIDER_CAPABILITIES
-  );
-}
-
-export function isBuiltInSearchProviderType(
-  providerType: string
-): providerType is WebSearchProviderType {
-  return Object.prototype.hasOwnProperty.call(
-    SEARCH_PROVIDER_DETAILS,
-    providerType
-  );
-}
-
-export function searchProviderRequiresApiKey(providerType: string): boolean {
-  return getCapabilities(providerType).requiresApiKey;
-}
-
-function getStoredConfigValue(
-  providerType: string,
-  canonicalKey: string,
-  config: SearchProviderConfig
-): string {
-  const caps = getCapabilities(providerType);
-  const aliases = caps.storedConfigAliases?.[canonicalKey] ?? [canonicalKey];
-
-  const safeConfig = config ?? {};
-  for (const key of aliases) {
-    const value = safeConfig[key];
-    if (typeof value === "string" && value.length > 0) {
-      return value;
-    }
-  }
-  return "";
-}
-
-/** True when the provider has all required credentials/config to be usable. */
-export function isSearchProviderConfigured(
-  providerType: string,
-  provider: SearchProviderLike
-): boolean {
-  const caps = getCapabilities(providerType);
-
-  if (caps.requiresApiKey && !(provider?.has_api_key ?? false)) {
-    return false;
-  }
-
-  for (const requiredKey of caps.requiredConfigKeys) {
-    const value = getStoredConfigValue(
-      providerType,
-      requiredKey,
-      provider?.config
-    );
-    if (!value) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-export function canConnectSearchProvider(
-  providerType: string,
-  apiKey: string,
-  searchEngineIdOrBaseUrl: string
-): boolean {
-  const caps = getCapabilities(providerType);
-
-  if (caps.requiresApiKey && apiKey.trim().length === 0) {
-    return false;
-  }
-
-  // Today, all config-driven search providers only expose a single required string field.
-  if (
-    caps.requiredConfigKeys.length > 0 &&
-    searchEngineIdOrBaseUrl.trim().length === 0
-  ) {
-    return false;
-  }
-
-  return true;
-}
-
-/** Build the `config` payload to send to the backend for a provider. */
-export function buildSearchProviderConfig(
-  providerType: string,
-  searchEngineIdOrBaseUrl: string
-): Record<string, string> {
-  const caps = getCapabilities(providerType);
-  const value = searchEngineIdOrBaseUrl.trim();
-
-  const config: Record<string, string> = {};
-  if (!value || caps.requiredConfigKeys.length === 0) {
-    return config;
-  }
-
-  // Only one required key for now.
-  const requiredKey = caps.requiredConfigKeys[0];
-  if (!requiredKey) {
-    return config;
-  }
-  config[requiredKey] = value;
-  return config;
-}
-
-/**
- * For providers that have a single required config field, return that stored value for form prefilling.
- */
-export function getSingleConfigFieldValueForForm(
-  providerType: string,
-  provider: SearchProviderLike
-): string {
-  const caps = getCapabilities(providerType);
-  if (caps.requiredConfigKeys.length === 0) {
-    return "";
-  }
-
-  const requiredKey = caps.requiredConfigKeys[0];
-  if (!requiredKey) {
-    return "";
-  }
-  return getStoredConfigValue(providerType, requiredKey, provider?.config);
-}
